refactor(hir): split for-await target-detection helpers out of body_stmt.rs (file-size gate)
body_stmt.rs hit 2038 lines on main (limit 2000), so the lint job's
file-size gate fails every PR even after the fmt fix. Move the
self-contained for-await/for-of head-expression detection predicates
(ReadableStream / Node Readable / readline / fs.Dir) plus the
IteratorClose helpers into body_stmt/detect.rs, following the existing
body_stmt/for_await.rs split pattern. Pure code motion, no behavior
change; body_stmt.rs is now 1826 lines.
